+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),  
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2025-02-25
+
+### Added
+- A "Manage Search Engines" option was added as a child of the "Contextual" top-level right click menu for quick access
+- An initial default state is now provided 
+
 ## [1.0.0] - 2025-02-24
 
 ### Added